Created user section Added roles Added restrictions Fixed email verification Fixed table names Fixed axios error handler

+++ blog/Http/Controllers/Api/Admin/CategoryController.php
@@ -29,12 +29,19 @@ class CategoryController extends Controller
 
     public function store(Request $request)
     {
+        if (!in_array(auth()->user()->id, [1,3,5]) and
+            Category::where('created_by', auth()->user()->id)->count() > 1
+        ) {
+            abort(403, 'You can create only 2 categories');
+        }
+
         $rules = $this->rules;
-        $rules['name'][] = 'unique:categories,name';
+        $rules['name'][] = 'unique:blog_categories,name';
         $request->validate($rules);
 
         $category = new Category();
         $category->fill($request->all());
+        $category->createdBy()->associate(auth()->user());
         $category->save();
 
         return response()->json($category);
@@ -42,8 +49,14 @@ class CategoryController extends Controller
 
     public function update(Request $request, Category $category)
     {
+        if (!in_array(auth()->user()->id, [1,3,5]) and
+            $category->createdBy->id != auth()->user()->id
+        ) {
+            abort(403, 'You can edit only your categories');
+        }
+
         $rules = $this->rules;
-        $rules['name'][] = "unique:categories,name,{$category->id}";
+        $rules['name'][] = "unique:blog_categories,name,{$category->id}";
         $request->validate($rules);
 
         $category->fill($request->all());
@@ -54,6 +67,12 @@ class CategoryController extends Controller
 
     public function destroy(Category $category)
     {
+        if (!in_array(auth()->user()->id, [1,3,5]) and
+            $category->createdBy->id != auth()->user()->id
+        ) {
+            abort(403, 'You can delete only your categories');
+        }
+
         $category->delete();
 
         return response()->json();
