+++ blog/Http/Controllers/Api/Admin/PostController.php
@@ -45,8 +45,14 @@ class PostController extends Controller
 
     public function store(Request $request)
     {
+        if (!in_array(auth()->user()->id, [1,3,5]) and
+            Post::where('author_id', auth()->user()->id)->count() > 1
+        ) {
+            abort(403, 'You can create only 2 posts');
+        }
+
         $rules = $this->rules;
-        $rules['title'][] = 'unique:posts,title';
+        $rules['title'][] = 'unique:blog_posts,title';
         $request->validate($rules);
 
         $post = new Post();
@@ -59,8 +65,14 @@ class PostController extends Controller
 
     public function update(Request $request, Post $post)
     {
+        if (!in_array(auth()->user()->id, [1,3,5]) and
+            $post->author->id != auth()->user()->id
+        ) {
+            abort(403, 'You can edit only your posts');
+        }
+
         $rules = $this->rules;
-        $rules['title'][] = "unique:posts,title,{$post->id}";
+        $rules['title'][] = "unique:blog_posts,title,{$post->id}";
         $request->validate($rules);
 
         $this->save($request, $post);
@@ -70,6 +82,12 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        if (!in_array(auth()->user()->id, [1,3,5]) and
+            $post->author->id != auth()->user()->id
+        ) {
+            abort(403, 'You can delete only your posts');
+        }
+
         $post->delete();
 
         return response()->json();
